Adapt e2e tests for remote instance (#4210)
* adapt photofinish and calculate wanda url

* fix prettier issue

* adapt hana cluster details test suite for remote target

* set api key and auto discover api key env variables for cypress

* fixes for agent heartbeat function

* fix apideregister instance

* move get api key logic to cypress config & fix startAgentHeartbeat to work with remote instances

* remove ignore uncaught exceptions

* use common mock data for suma

* revert changes on elixir suma mock and refresh api key in after hook in settings tests

* apply linter to sw updates mocks

* remove auto discover api key and fetch it always before photofinish to ensure it will work

* fixess

* variable naming fixes

* increase timeouts hosts overview

* fixes on start agents heartbeat for remote host

* restore hosts

* remove tls reject unauthorized

* ignore uncaught for sles16 target

* conditional validation of subscriptions & api cleanup when target host is a real systemd instance

* add explanatory comments

* fix format on e2e.js

* polish details & add logic to not get apikey when starting heartbeat against localhost

* add missing return statements for cypress to chain commands properly

* add missing return statements

* add missing return statements for cypress chaining and fix loadscenario to properly return too

* wait for host to be restored

* relogin in sap_systems_overview due to login expiration and increase timeout in host etails

* assert host is restored before

* visit after restore

* host details tests more robust

* add retry for restore host in host details view when running against remote instance

* fix race condition issues on sap systems overview

* remove unused wrapper

* explicitly set wanda & web variables to indicate if the instance is a systemd one or dev

* fix linting

* returns refactor

* fix conflict and run lint

* set default timeout to 10s when waiting for requests

* set lint rules for using arrow-body-style and prefer-arrow-callback, also apply them.

* prettier

* address pr feedback

* run prettier

* prettier & cy.wrap where required

* remove conditionals for suse manager settings calls in before hook

* format

* remove hack-ish refreshLogin and waitForRestoreHost

* use demo and prod values for web and wanda mode

* fix trigger email

* reword real to prod

* some more fixes and remove unused code

* calculate global default timeouts based on web_mode env param

* lint format

* logic for loadScenario timeout

* remove logic for commands timeout

+++ test/e2e/cypress/pageObject/alerting_po.js
@@ -40,7 +40,7 @@ export const emailIsReceived = (type) =>
     .then((result) => cy.wrap(result.length).should('equal', 1));
 
 export const triggerHostAlertingEmail = () => {
-  cy.task('startAgentHeartbeat', ['9cd46919-5f19-59aa-993e-cf3736c71053']);
+  basePage.startAgentsHeartbeat(['9cd46919-5f19-59aa-993e-cf3736c71053']);
   return basePage.stopAgentsHeartbeat();
 };
 
@@ -55,5 +55,6 @@ export const triggerSapSystemAlertingEmail = () =>
 export const triggerDatabaseAlertingEmail = () =>
   basePage.loadScenario('hana-database-detail-RED');
 
-export const deleteAllEmailsFromMailpit = () =>
-  cy.task('deleteAllEmailsFromMailpit');
+export const deleteAllEmailsFromMailpit = () => {
+  if (Cypress.env('ALERTING_TESTS')) cy.task('deleteAllEmailsFromMailpit');
+};
